# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Brgy.Tanod S.O.S Emergency Response API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: residentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create SOS alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get alert by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found

    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/respond:
    post:
      operationId: respondToAlert
      tags: [alerts]
      summary: Tanod responds to alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAlertInput"
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/cancel:
    post:
      operationId: cancelAlert
      tags: [alerts]
      summary: Cancel alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/messages:
    get:
      operationId: listAlertMessages
      tags: [messages]
      summary: Get chat messages for alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertMessage"

    post:
      operationId: createAlertMessage
      tags: [messages]
      summary: Send message in alert chat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertMessage"
...
  /broadcasts:
    get:
      operationId: listBroadcasts
      tags: [broadcasts]
      summary: List system broadcasts
      parameters:
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Broadcasts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Broadcast"

    post:
      operationId: createBroadcast
      tags: [broadcasts]
      summary: Create system broadcast (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "201":
          description: Broadcast created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
  /broadcasts/{id}:
    patch:
      operationId: updateBroadcast
      tags: [broadcasts]
      summary: Update broadcast
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastUpdate"
      responses:
        "200":
          description: Updated broadcast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [resident, tanod]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [resident, tanod, admin, superadmin]
        status:
          type: string
          enum: [pending, approved, rejected, active]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        barangayId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: ["string", "null"]
          format: date-time
...
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
        accuracy:
          type: ["number", "null"]
    Alert:
      type: object
      required: [id, type, status, location, createdAt]
      properties:
        id:
          type: string
        residentId:
          type: ["string", "null"]
        residentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [MEDICAL, FIRE, CRIME, NATURAL_DISASTER, VIOLENCE, FLOOD, DISTURBANCE, OTHER]
        status:
          type: string
          enum: [pending, responding, resolved, cancelled]
        location:
          $ref: "#/components/schemas/Location"
        description:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        assignedToName:
          type: ["string", "null"]
        respondedBy:
          type: ["string", "null"]
        respondedByName:
          type: ["string", "null"]
        respondedAt:
          type: ["string", "null"]
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        resolutionNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AlertInput:
      type: object
      required: [type, location]
      properties:
        type:
          type: string
          enum: [MEDICAL, FIRE, CRIME, NATURAL_DISASTER, VIOLENCE, FLOOD, DISTURBANCE, OTHER]
        location:
          $ref: "#/components/schemas/Location"
        description:
          type: ["string", "null"]
    AlertUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        assignedToName:
          type: ["string", "null"]
        resolutionNotes:
          type: ["string", "null"]
        responderNotes:
          type: ["string", "null"]
    ResolveAlertInput:
      type: object
      properties:
        resolutionNotes:
          type: ["string", "null"]
    AlertMessage:
      type: object
      required: [id, alertId, senderId, senderName, senderRole, message, timestamp]
      properties:
        id:
          type: string
        alertId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderRole:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    AlertMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Broadcast:
      type: object
      required: [id, message, timestamp, isActive]
      properties:
        id:
          type: string
        incidentId:
          type: ["string", "null"]
        adminId:
          type: ["string", "null"]
        adminName:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [evacuation, calamity, security, emergency, other, null]
        message:
          type: string
        isActive:
          type: boolean
        approvalStatus:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
    BroadcastInput:
      type: object
      required: [message, type]
      properties:
        message:
          type: string
        type:
          type: string
          enum: [evacuation, calamity, security, emergency, other]
        incidentId:
          type: ["string", "null"]
    BroadcastUpdate:
      type: object
      properties:
        isActive:
          type: ["boolean", "null"]
        approvalStatus:
          type: ["string", "null"]
...