# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brgy.Tanod S.O.S Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertStatus } from './alertStatus';
import type { AlertType } from './alertType';
import type { Location } from './location';

export interface Alert {
  id: string;
  /** @nullable */
  residentId?: string | null;
  /** @nullable */
  residentName?: string | null;
  type: AlertType;
  status: AlertStatus;
  location: Location;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  assignedToName?: string | null;
  /** @nullable */
  respondedBy?: string | null;
  /** @nullable */
  respondedByName?: string | null;
  /** @nullable */
  respondedAt?: Date | null;
  /** @nullable */
  resolvedAt?: Date | null;
  /** @nullable */
  resolutionNotes?: string | null;
  createdAt: Date;
}